pyproject.toml
[project]
name = "tachyon-model"
version = "0.1.0"
description = "Config-driven downloader for Chronos base models."
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
  "huggingface_hub>=0.24.0",
  "pydantic>=2.7.0",
  "pyyaml>=6.0.0",
  "autogluon-timeseries>=1.1.0",
  "pandas>=2.2.0",
  "pyarrow>=15.0.0",
  "chronos-forecasting>=1.4.0",
  "torch>=2.2.0",
]

[tool.poetry]
packages = [{ include = "tachyon_model_downloader", from = "src" }]

[tool.poetry.group.dev.dependencies]
ruff = ">=0.11.0"
mypy = ">=1.11.0"
pre-commit = ">=3.8.0"
bandit = ">=1.7.0"

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "W"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
files = ["src"]

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
